+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## TBD
+### Changed
+- MessageReader to return MessageEvent which wraps a ListenerEvent and MessageProperty.
+
+### Added
+- `SqsMessageReader.delete()` and `SqsMessageReader.extendVisibilityTimeout()` methods.
+
 ### Removed
 - `aws-java-sdk` dependency from `apiary-receiver-sqs`.